refactor: split post-cutover evidence repository
