+++ src/backend/src/agentclaw/community/plugins/skills_pool_layout_repository.py
@@ -32,6 +32,9 @@ from agentclaw.community.plugins.skills_pool_capability_repository import (
 from agentclaw.community.plugins.skills_pool_operational_repository import (
     SkillsPoolOperationalRepositoryMixin,
 )
+from agentclaw.community.plugins.skills_pool_post_cutover_repository import (
+    SkillsPoolPostCutoverRepositoryMixin,
+)
 from agentclaw.community.plugins.skills_pool_quarantine_repository import (
     SkillsPoolQuarantineRepositoryMixin,
 )
@@ -40,6 +43,7 @@ from agentclaw.community.plugins.skills_pool_quarantine_repository import (
 class SkillsPoolLayoutRepository(
     SkillsPoolCapabilityRepositoryMixin,
     SkillsPoolOperationalRepositoryMixin,
+    SkillsPoolPostCutoverRepositoryMixin,
     SkillsPoolQuarantineRepositoryMixin,
 ):
     @inject
@@ -407,68 +411,6 @@ class SkillsPoolLayoutRepository(
             row.last_failure_retryable = 1
             row.last_failure_evidence = evidence_json
             row.last_failure_at = func.now()
-        return True
-
-    def record_post_cutover_evidence(
-        self,
-        *,
-        scope: BotSkillLayoutScope,
-        migration_generation: str,
-        lease_owner: str,
-        preparation_id: str,
-        evidence: dict[str, object],
-    ) -> bool:
-        """Reconcile runtime evidence without re-crossing the cutover boundary."""
-
-        evidence_json = json.dumps(evidence, ensure_ascii=False)
-        runtime_evidence = evidence.get("evidence")
-        quarantine_path = (
-            runtime_evidence.get("quarantine")
-            if isinstance(runtime_evidence, dict)
-            else None
-        )
-        with self._database.transactional_orm_session() as session:
-            row = (
-                session.query(BotSkillLayoutStateModel)
-                .filter(
-                    *self._scope_filter(scope),
-                    BotSkillLayoutStateModel.active_layout == SkillLayout.LEGACY.value,
-                    BotSkillLayoutStateModel.target_layout == SkillLayout.POOL.value,
-                    BotSkillLayoutStateModel.phase.in_(
-                        (
-                            SkillLayoutPhase.POOL_CUTOVER_FINALIZING.value,
-                            SkillLayoutPhase.POOL_CUTOVER_COMMITTED.value,
-                        )
-                    ),
-                    BotSkillLayoutStateModel.data_plane_cutover_committed == 1,
-                    BotSkillLayoutStateModel.migration_generation
-                    == migration_generation,
-                    BotSkillLayoutStateModel.preparation_id == preparation_id,
-                    BotSkillLayoutStateModel.lease_owner == lease_owner,
-                    BotSkillLayoutStateModel.lease_expires_at > func.now(),
-                )
-                .with_for_update()
-                .one_or_none()
-            )
-            if row is None or row.rollout_evidence is None:
-                return False
-            engine = json.loads(row.rollout_evidence).get("engine_type")
-            if not isinstance(engine, str) or not engine:
-                return False
-            if not isinstance(quarantine_path, str) or not quarantine_path:
-                log_missing_quarantine_path(scope, migration_generation)
-                return False
-            if not self._upsert_quarantine(
-                session,
-                scope=scope,
-                migration_generation=migration_generation,
-                engine=engine,
-                path=quarantine_path,
-                evidence_json=evidence_json,
-            ):
-                return False
-            row.phase = SkillLayoutPhase.POOL_CUTOVER_COMMITTED.value
-            row.last_probe_evidence = evidence_json
         return True
 
     def begin_cutover(
